requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.0
frozenlist==1.3.1
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.8.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
multidict==6.0.2
orjson==3.7.11
Pillow==9.2.0
py-cord==2.0.0
pycparser==2.21
python-dotenv==0.20.0
rfc3986==1.5.0
sniffio==1.2.0
yarl==1.8.1
